Deduplicate and clean up examples
