+++ cpp/examples/parquet_io/CMakeLists.txt
@@ -23,7 +23,7 @@ rapids_cmake_build_type("Release")
 # gcc>=14. We can remove this once we upgrade to a newer sccache version.
 set(CMAKE_CXX_SCAN_FOR_MODULES OFF)
 
-add_library(parquet_io_utils OBJECT common_utils.cpp io_source.cpp)
+add_library(parquet_io_utils OBJECT parquet_io_utils.cpp)
 target_compile_features(parquet_io_utils PRIVATE cxx_std_20)
 target_link_libraries(parquet_io_utils PRIVATE cudf::cudf)
