+++ cpp/examples/utilities/timer.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024-2025, NVIDIA CORPORATION.
+ * Copyright (c) 2025, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,8 +19,8 @@
 #include <chrono>
 #include <iostream>
 
-namespace cudf {
-namespace examples {
+namespace cudf::examples {
+
 /**
  * @brief Light-weight timer for measuring elapsed time.
  *
@@ -52,5 +52,4 @@ class timer {
   time_point_t start_time;
 };
 
-}  // namespace examples
-};  // namespace cudf
+}  // namespace cudf::examples
